v1.01
Added an unloader, reloader, and got most of the selection based on
character race set up.  Just need to get it to load the icons now.  I
think it's getting much closer.  It's better than the two plugins it's
based on that won't even load giving an error of "Unable to resolve
package "RogidorPlugins.MountCarouselXX.Main".".